requirements.txt
attrs==20.2.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.1
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.60.2
fastapi-users==3.0.3
fastapi-utils==0.2.1
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
makefun==1.9.2
mongomock==3.20.0
motor==2.1.0
packaging==20.4
passlib==1.7.2
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.0
python-dotenv==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
requests==2.24.0
rsa==4.6
sentinels==1.0.0
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
toml==0.10.1
typing==3.7.4.3
urllib3==1.25.10
uuid==1.30
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
zipp==3.2.0
